app/src/main/java/com/esoxjem/movieguide/listing/MoviesListingPresenter.java
package com.esoxjem.movieguide.listing;

import com.esoxjem.movieguide.Movie;
import com.esoxjem.movieguide.util.RxUtils;

import java.util.List;

import rx.Subscriber;
import rx.Subscription;
import rx.android.schedulers.AndroidSchedulers;
import rx.schedulers.Schedulers;

/**
 * @author arun
 */
public class MoviesListingPresenter implements IMoviesListingPresenter
{
    private IMoviesListingView view;
    private IMoviesListingInteractor moviesInteractor;
    private Subscription fetchSubscription;

    public MoviesListingPresenter(IMoviesListingInteractor interactor)
    {
        moviesInteractor = interactor;
    }

    @Override
    public void setView(IMoviesListingView view)
    {
        this.view = view;
    }

    @Override
    public void destroy()
    {
        view = null;
        RxUtils.unsubscribe(fetchSubscription);
    }

    @Override
    public void displayMovies()
    {
        showLoading();
        fetchSubscription = moviesInteractor.fetchMovies().subscribeOn(Schedulers.io())
                .observeOn(AndroidSchedulers.mainThread())
                .subscribe(new Subscriber<List<Movie>>()
                {
                    @Override
                    public void onCompleted()
                    {

                    }

                    @Override
                    public void onError(Throwable e)
                    {
                        onMovieFetchFailed(e);
                    }

                    @Override
                    public void onNext(List<Movie> movies)
                    {
                        onMovieFetchSuccess(movies);
                    }
                });
    }

    void showLoading()
    {
        if (isViewAttached())
        {
            view.loadingStarted();
        }
    }

    void onMovieFetchSuccess(List<Movie> movies)
    {
        view.showMovies(movies);
    }

    void onMovieFetchFailed(Throwable e)
    {
        view.loadingFailed(e.getMessage());
    }

    private boolean isViewAttached()
    {
        return view != null;
    }
}
